styles: 반응형 수정및 컴포넌트 통합
