+++ src/assets/scss/Styles.scss
@@ -32,5 +32,3 @@
 @import './ThumbnailList/ThumbnailList.scss';
 // 내쇼핑몰 scss
 @import '../scss/Myshop.scss';
-@import '../scss/ShopImg.scss';
-@import '../scss/ShopInfo.scss';
