+++ src/pages/Edit.jsx
@@ -1,5 +1,13 @@
+import RepItem from '../components/common/RepItem'
+import MyShop from '../components/common/Myshop'
+
 const Edit = () => {
-  return <div>Edit Page</div>
+  return (
+    <div>
+      <RepItem />
+      <MyShop />
+    </div>
+  )
 }
 
 export default Edit
